Create atlas.hcl
The Atlas configuration file is necessary to lint SQL files that contain
Go template syntax. You can now run the lint command with
`atlas migrate lint \
--dev-url 'docker://postgres/15/dev?search_path=public'
--env default \
--var env=stg \
--latest 10`
This command should warn you that the custom migration file drops the
users table.
Note that if you run the above command with any other value for
`--var env=`, the linter will not complain, as it will expand the Go
template to be a blank string in that case.